Move availability to calendars resource
